+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules/
 docs/
-!docs/conf.json
 examples/
 build/
