--- docs/conf.json
@@ -1,14 +1,0 @@
-{
-    "tags": {
-        "allowUnknownTags": true,
-        "dictionaries": ["jsdoc", "closure"]
-    },
-    "source": {
-        "include": ["src/"]
-    },
-    "plugins": [],
-    "templates": {
-        "cleverLinks": false,
-        "monospaceLinks": false
-    }
-}
