+++ src/HyperlapseView.js
@@ -1,0 +1,5 @@
+var HyperlapseView = function(){
+  
+};
+
+module.exports = HyperlapseView;
